+++ docs/changelog.rst
@@ -11,6 +11,8 @@ Changelog
 - Rename the ``normalized_parameter_values`` keyword of
   :func:`headers._parse_parameter_list`.  The current spelling is retained
   with a deprecation warning.  This will be removed in 2.0.
+- Add ``normalize_parameter_names`` keyword to the
+  :func:`headers._parse_parameter_list` internal function.
 
 `1.4.3`_ (30-Oct-2017)
 ----------------------